Improved connectivity checks
Used the -w argument on curl to get the http return code and compare it to something of type [1-9][0-9][0-9]. If the website is not reachable, it should return 000 not matching the regex. If the website responds but is something else than 200, it should work. If return code is 407 (proxy authentication request) or 502 (proxy error) it should mention it.